Exception handling - delete

+++ src/main/java/com/estudos/services/UserService.java
@@ -4,10 +4,13 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.estudos.entities.User;
 import com.estudos.repositories.UserRepository;
+import com.estudos.services.exceptions.DatabaseException;
 import com.estudos.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -30,9 +33,17 @@ public class UserService {
 	}
 	
 	public void deleteById(Long id) {
-		repository.deleteById(id);
+	    if (!repository.existsById(id)) {
+	        throw new ResourceNotFoundException(id);
+	    }
+	    try {
+	    repository.deleteById(id);
+	    }
+	    catch(DataIntegrityViolationException e) {
+	     throw new DatabaseException(e);
+	    }
 	}
-	
+
 	public User update(Long id, User obj) {
 		User entity = repository.getReferenceById(id);
 		updateData(entity, obj);
